+++ src/lib.rs
@@ -1,3 +1,5 @@
+//! TODO: Write module documentation
+
 #[cfg(test)]
 mod test;
 
@@ -9,6 +11,8 @@ use model::{
     AnimeDetails, AnimeList, ForumBoards, ForumTopics, ListStatus, TopicDetails, User,
 };
 
+use aes_gcm::aead::{Aead, NewAead};
+use aes_gcm::{Aes256Gcm, Key, Nonce};
 use pkce;
 use rand::random;
 use reqwest::{Method, StatusCode};
@@ -16,6 +20,7 @@ use serde::{Deserialize, Serialize};
 use std::{
     fs::{self, File},
     io::Write,
+    path::Path,
     path::PathBuf,
     str,
     time::SystemTime,
@@ -40,6 +45,7 @@ impl MALClient {
         let client = reqwest::Client::new();
         let mut will_cache = caching;
         let mut n_a = false;
+
         let dir = if let Some(d) = cache_dir {
             d
         } else {
@@ -47,10 +53,12 @@ impl MALClient {
             will_cache = false;
             PathBuf::new()
         };
+
+        println!("Current path exists: {}", dir.join("tokens").exists());
         let mut token = String::new();
-        if will_cache && dir.join("tokens.json").exists() {
-            if let Ok(tokens) = fs::read_to_string(dir.join("tokens.json")) {
-                let mut tok: Tokens = serde_json::from_str(&tokens).unwrap();
+        if will_cache && dir.join("tokens").exists() {
+            if let Ok(tokens) = fs::read(dir.join("tokens")) {
+                let mut tok: Tokens = decrypt_tokens(&tokens).unwrap();
                 if let Ok(n) = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH) {
                     if n.as_secs() - tok.today >= tok.expires_in as u64 {
                         let params = [
@@ -77,18 +85,15 @@ impl MALClient {
                                 .as_secs(),
                         };
 
-                        fs::write(
-                            dir.join("tokens.json"),
-                            serde_json::to_string(&tok).expect("Unable to parse token struct"),
-                        )
-                        .expect("Unable to write token file")
+                        fs::write(dir.join("tokens"), encrypt_token(tok).unwrap())
+                            .expect("Unable to write token file")
                     } else {
                         token = tok.access_token;
                     }
                 }
             }
         } else {
-            will_cache = false;
+            will_cache = caching;
             n_a = true;
         }
 
@@ -100,47 +105,49 @@ impl MALClient {
             client,
             caching: will_cache,
         };
-
         me
     }
 
-    ///Returns the auth URL and code challenge which will be needed to authorize the user
+
+    ///Returns the auth URL and code challenge which will be needed to authorize the user.
+    ///
+    ///# Example
     ///
     ///```rust
-    /// use lib_mal::MALClient;
-    /// use tokio
+    ///     use lib_mal::MALClient;
     ///
-    /// #[tokio::main]
-    /// async fn main() {
-    ///     let redirect_uri = [YOUR_REDIRECT_URI_HERE];
+    ///     let redirect_uri = "http://localhost:2525";//<-- example uri
     ///     let client = MALClient::new([YOUR_SECRET_HERE]).await;
-    ///     let (url, challenge, state) = client.get_auth_parts(&redirect_uri);
+    ///     let (url, challenge, state) = client.get_auth_parts();
     ///     println!("Go here to log in: {}", url);
     ///     client.auth(&redirect_uri, &challenge, &state).await.expect("Unable to log in");
-    ///     //Auth will open an http server on port 2561 to listen for the OAuth2 callback
     /// }
     ///```
-    pub fn get_auth_parts(&self, callback_url: &str) -> (String, String, String) {
+    pub fn get_auth_parts(&self) -> (String, String, String) {
         let verifier = pkce::code_verifier(128);
         let challenge = pkce::code_challenge(&verifier);
         let state = format!("bruh{}", random::<u8>());
-        let url = format!("https://myanimelist.net/v1/oauth2/authorize?response_type=code&client_id={}&code_challenge={}&redirect_uri={}&state={}", self.client_secret, challenge, callback_url, state);
+        let url = format!("https://myanimelist.net/v1/oauth2/authorize?response_type=code&client_id={}&code_challenge={}&state={}", self.client_secret, challenge, state, );
         (url, challenge, state)
     }
 
-    ///Listens for the OAuth2 callback from MAL on `callback_url`, which is the callback url
-    ///registered when obtaining the API token from MAL. Only applications with a single registered
-    ///URL are supported at the moment.
+    ///Listens for the OAuth2 callback from MAL on `callback_url`, which is the redirect_uri
+    ///registered when obtaining the API token from MAL. Only HTTP URIs are supported right now.
+    ///
+    ///For now only applications with a single registered URI are supported, having more than one
+    ///seems to cause issues with the MAL api itself
+    ///
+    ///# Example
     ///
     ///```rust
-    /// #[tokio::main]
-    /// async fn main() {
-    ///     let redirect_uri = [YOUR_REDIRECT_URI_HERE];
+    ///     use lib_mal::MALClient;
+    ///
+    ///     let redirect_uri = "localhost:2525";//<-- example uri,
+    ///     //appears as "http://localhost:2525" in the API settings
     ///     let client = MALClient::new([YOUR_SECRET_HERE]).await;
     ///     let (url, challenge, state) = client.get_auth_parts(&redirect_uri);
     ///     println!("Go here to log in: {}", url);
     ///     client.auth(&redirect_uri, &challenge, &state).await.expect("Unable to log in");
-    ///     //Auth will open an http server on port 2561 to listen for the OAuth2 callback
     /// }
     ///```
     pub async fn auth(
@@ -150,8 +157,17 @@ impl MALClient {
         state: &str,
     ) -> Result<(), String> {
         let mut code = "".to_owned();
+        let url = if callback_url.contains("http") {
+            //server won't work if the url has the protocol in it
+            callback_url
+                .trim_start_matches("http://")
+                .trim_start_matches("https://")
+        } else {
+            callback_url
+        };
 
-        let server = Server::http(callback_url).unwrap();
+
+        let server = Server::http(url).unwrap();
         for i in server.incoming_requests() {
             if !i.url().contains(&format!("state={}", state)) {
                 //if the state doesn't match, discard this response
@@ -171,11 +187,10 @@ impl MALClient {
             break;
         }
 
-        self.get_tokens(&code, &challenge).await;
-        Ok(())
+        self.get_tokens(&code, &challenge).await
     }
 
-    async fn get_tokens(&mut self, code: &str, verifier: &str) {
+    async fn get_tokens(&mut self, code: &str, verifier: &str) -> Result<(), String> {
         let params = [
             ("client_id", self.client_secret.as_str()),
             ("grant_type", "authorization_code"),
@@ -189,23 +204,28 @@ impl MALClient {
             .build()
             .unwrap();
         let res = self.client.execute(rec).await.unwrap();
-        let tokens: TokenResponse = serde_json::from_str(&res.text().await.unwrap()).unwrap();
-        self.access_token = tokens.access_token.clone();
+        let text = res.text().await.unwrap();
+        if let Ok(tokens) = serde_json::from_str::<TokenResponse>(&text) {
+            self.access_token = tokens.access_token.clone();
 
-        let tjson = Tokens {
-            access_token: tokens.access_token,
-            refresh_token: tokens.refresh_token,
-            expires_in: tokens.expires_in,
-            today: SystemTime::now()
-                .duration_since(SystemTime::UNIX_EPOCH)
-                .unwrap()
-                .as_secs(),
-        };
-        if self.caching {
-            let mut f =
-                File::create(self.dirs.join("tokens.json")).expect("Unable to create token file");
-            f.write_all(serde_json::to_string(&tjson).unwrap().as_bytes())
-                .expect("Unable to write tokens");
+            let tjson = Tokens {
+                access_token: tokens.access_token,
+                refresh_token: tokens.refresh_token,
+                expires_in: tokens.expires_in,
+                today: SystemTime::now()
+                    .duration_since(SystemTime::UNIX_EPOCH)
+                    .unwrap()
+                    .as_secs(),
+            };
+            if self.caching {
+                let mut f =
+                    File::create(self.dirs.join("tokens")).expect("Unable to create token file");
+                f.write_all(&encrypt_token(tjson).unwrap())
+                    .expect("Unable to write tokens");
+            }
+            Ok(())
+        }else {
+            Err(text) 
         }
     }
 
@@ -473,6 +493,26 @@ impl MALClient {
         let res = self.do_request(url).await?;
         self.parse_response(&res)
     }
+}
+
+fn encrypt_token(toks: Tokens) -> Result<Vec<u8>, ()> {
+    let key = Key::from_slice(b"one two three four five six seve");
+    let cypher = Aes256Gcm::new(&key);
+    let nonce = Nonce::from_slice(b"but the eart");
+    let plain = serde_json::to_vec(&toks).unwrap();
+    let res = cypher.encrypt(nonce, plain.as_ref()).unwrap();
+    Ok(res)
+}
+
+fn decrypt_tokens(raw: &Vec<u8>) -> Result<Tokens, ()> {
+    let key = Key::from_slice(b"one two three four five six seve");
+    let cypher = Aes256Gcm::new(&key);
+    let nonce = Nonce::from_slice(b"but the eart");
+    let plain = cypher
+        .decrypt(nonce, raw.as_ref())
+        .expect("Couldn't decrypt");
+    let text = String::from_utf8(plain).unwrap();
+    Ok(serde_json::from_str(&text).expect("couldn't parse decrypted tokens"))
 }
 
 #[derive(Deserialize, Debug)]